Converted custom IDS of feeds/categories to QString from int - probably broken.

+++ src/services/owncloud/gui/formowncloudfeeddetails.cpp
@@ -68,7 +68,7 @@ void FormOwnCloudFeedDetails::apply() {
   else {
     const RootItem* parent = static_cast<RootItem*>(m_ui->m_cmbParentCategory->itemData(
                                                       m_ui->m_cmbParentCategory->currentIndex()).value<void*>());
-    const int category_id = parent->kind() == RootItemKind::ServiceRoot ? 0 : parent->customId();
+    const int category_id = parent->kind() == RootItemKind::ServiceRoot ? 0 : parent->customId().toInt();
     const bool response = qobject_cast<OwnCloudServiceRoot*>(m_serviceRoot)->network()->createFeed(m_ui->m_txtUrl->lineEdit()->text(),
                                                                                                    category_id);
